Remove old implementation and also implement missing variables and methods.
